- [GLSL]Merge all ray tracing code to GLSL 	* Change double to float to prevent any type issues 	* Add 'Surface' structure to describe inheritance with all surfaces 	* New 'General' uniform
- [Data setup]Data write to buffer and read back in app without issues
	* New #define for structure size and enum for types
	* Changes in 'TraiterFichierScene' to get the index of the surface in its array
	* Solve many issues with data copy to buffer
	* Solve issue with texture assignement and binding
	* After dispatch, display data in buffer
- [Other] Change REAL type from 'float' to 'double' (to match GLSL requirements)
- [Other]Methods to fill and print 'Surface' structure
- [Other]New methods 'AjusterNormale' and 'ObtenirNormale' in Triangle

+++ Projet-INF8702/MathUtils.h
@@ -18,7 +18,12 @@
 namespace Math3D
 {
     ///////////////////// TYPEDEFS ///////////////////////
-    typedef double REAL;
+    
+	//TODO : check if no drawbacks
+	//new
+	typedef float REAL;
+	//old
+	//typedef double REAL;
 
     ///////////////////// DEFINES ////////////////////////
     #ifndef M_PI
